code refactor, folder structure, better bg, loading screen, about
