+++ index.html
@@ -7,11 +7,20 @@
         <link rel="stylesheet" type="text/css" href="index.css">
     </head>
 
+    <!-- author: github.com/Z04P -->
     <!-- built using PixiJS v4.5.1 and sound.js by kittykatattack on github -->
+    <!-- frameworks/libraries -->
     <script src="scripts/pixi.min.js"></script>
     <script src="scripts/sound.js"></script>
 
+    <!-- app scripts -->
+    <script src="scripts/load-stuff.js"></script>
+    <script src="scripts/mouse-control.js"></script>
+    <script src="scripts/mechanics.js"></script>
+    <script src="scripts/states.js"></script>
+
     <body>
+        <!-- created a div container to center the app -->
         <div id="container"></div>
         <script type="text/javascript" src="app.js"></script>
     </body>
