--- resources/spritesheet/cloud.json
@@ -1,28 +1,0 @@
-{"frames": {
-
-"cloud000.png":
-{
-	"frame": {"x":0,"y":0,"w":500,"h":200},
-	"rotated": false,
-	"trimmed": false,
-	"spriteSourceSize": {"x":0,"y":0,"w":500,"h":200},
-	"sourceSize": {"w":500,"h":200}
-},
-"cloud001.png":
-{
-	"frame": {"x":500,"y":0,"w":500,"h":200},
-	"rotated": false,
-	"trimmed": false,
-	"spriteSourceSize": {"x":0,"y":0,"w":500,"h":200},
-	"sourceSize": {"w":500,"h":200}
-}},
-"meta": {
-	"app": "https://www.codeandweb.com/texturepacker",
-	"version": "1.0",
-	"image": "cloud.png",
-	"format": "RGBA8888",
-	"size": {"w":1000,"h":200},
-	"scale": "1",
-	"smartupdate": "$TexturePacker:SmartUpdate:3487f6c5ccdf042f0a375351236fd7f8:44a3757cff30d149db33d4eac75789f9:9ec8b6ce27a186caff8882d1507ee32e$"
-}
-}
