+++ sprite/restart-button.png
