+++ spritesheet/cloud.json
@@ -1,0 +1,76 @@
+{"frames": {
+
+"cloud000.png":
+{
+	"frame": {"x":0,"y":0,"w":500,"h":360},
+	"rotated": false,
+	"trimmed": false,
+	"spriteSourceSize": {"x":0,"y":0,"w":500,"h":360},
+	"sourceSize": {"w":500,"h":360}
+},
+"cloud001.png":
+{
+	"frame": {"x":500,"y":0,"w":500,"h":360},
+	"rotated": false,
+	"trimmed": false,
+	"spriteSourceSize": {"x":0,"y":0,"w":500,"h":360},
+	"sourceSize": {"w":500,"h":360}
+},
+"cloud002.png":
+{
+	"frame": {"x":1000,"y":0,"w":500,"h":360},
+	"rotated": false,
+	"trimmed": false,
+	"spriteSourceSize": {"x":0,"y":0,"w":500,"h":360},
+	"sourceSize": {"w":500,"h":360}
+},
+"cloud003.png":
+{
+	"frame": {"x":1500,"y":0,"w":500,"h":360},
+	"rotated": false,
+	"trimmed": false,
+	"spriteSourceSize": {"x":0,"y":0,"w":500,"h":360},
+	"sourceSize": {"w":500,"h":360}
+},
+"cloud004.png":
+{
+	"frame": {"x":0,"y":360,"w":500,"h":360},
+	"rotated": false,
+	"trimmed": false,
+	"spriteSourceSize": {"x":0,"y":0,"w":500,"h":360},
+	"sourceSize": {"w":500,"h":360}
+},
+"cloud005.png":
+{
+	"frame": {"x":500,"y":360,"w":500,"h":360},
+	"rotated": false,
+	"trimmed": false,
+	"spriteSourceSize": {"x":0,"y":0,"w":500,"h":360},
+	"sourceSize": {"w":500,"h":360}
+},
+"cloud006.png":
+{
+	"frame": {"x":1000,"y":360,"w":500,"h":360},
+	"rotated": false,
+	"trimmed": false,
+	"spriteSourceSize": {"x":0,"y":0,"w":500,"h":360},
+	"sourceSize": {"w":500,"h":360}
+},
+"cloud007.png":
+{
+	"frame": {"x":1500,"y":360,"w":500,"h":360},
+	"rotated": false,
+	"trimmed": false,
+	"spriteSourceSize": {"x":0,"y":0,"w":500,"h":360},
+	"sourceSize": {"w":500,"h":360}
+}},
+"meta": {
+	"app": "https://www.codeandweb.com/texturepacker",
+	"version": "1.0",
+	"image": "cloud.png",
+	"format": "RGBA8888",
+	"size": {"w":2000,"h":720},
+	"scale": "1",
+	"smartupdate": "$TexturePacker:SmartUpdate:454a200c398837d3eced1ad710b3be53:a0ef649ce336baedac125d7be5825aa4:9ec8b6ce27a186caff8882d1507ee32e$"
+}
+}
